pyproject.toml
[project]
name = "quantumedgedetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "pylatexenc>=2.10",
    "qiskit>=2.2.2",
    "qiskit-aer>=0.17.2",
    "requests>=2.32.5",
]
